# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Инспектор ИИ — API",
    "version": "1.0.0",
    "description": "Сверка проектной, рабочей и исполнительной документации по матрице из 132 параметров. Асинхронная pull-модель: загрузка возвращает process_id, результат забирается по нему."
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "inspector_session"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/documents/upload": {
      "post": {
        "summary": "Загрузить документы и получить process_id",
        "tags": [
          "api-v1"
        ],
        "description": "Пакет PDF/DOCX/XML (и чертежи DWG/DXF) с реестром файлов CSV/XLSX/JSON. Ответ сразу содержит process_id; проверка идёт асинхронно, статус — GET /api/v1/processes/{process_id}/status. С process_id — дозагрузка до финализации протокола.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "files"
                ],
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "registry": {
                    "type": "string",
                    "format": "binary"
                  },
                  "process_id": {
                    "type": "string",
                    "pattern": "^[0-9]+$"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Процесс создан или дополнен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "process_id": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "PENDING",
                        "PARSING",
                        "READY",
                        "VERIFYING",
                        "COMPLETED",
                        "FINALIZED",
                        "ERROR",
                        "CANCELLED"
                      ]
                    },
                    "upload_status": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      },
                      "description": "PD_UPLOADED | PD_PARTIAL | PD_MISSING и так же для RD, ID"
                    },
                    "scenario": {
                      "type": "string",
                      "enum": [
                        "FULL",
                        "PD_RD_ONLY",
                        "PD_ID_ONLY",
                        "RD_ID_ONLY",
                        "SINGLE_ONLY",
                        "PARTIALLY_LOADED",
                        "NO_DOCUMENTS"
                      ]
                    },
                    "accepted": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    },
                    "rejected": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    },
                    "notice": {
                      "type": "string"
                    }
                  },
                  "description": "Процесс создан или дополнен"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/processes": {
      "get": {
        "summary": "Процессы проверки",
        "tags": [
          "api-v1"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "object_id",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 100
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "summary": "Запустить проверку по уже загруженным документам",
        "tags": [
          "api-v1"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "object_id",
                  "document_ids"
                ],
                "properties": {
                  "object_id": {
                    "type": "string",
                    "minLength": 1
                  },
                  "document_ids": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                      "type": "integer",
                      "minimum": 1
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/processes/{process_id}/status": {
      "get": {
        "summary": "Мониторинг статуса процесса",
        "tags": [
          "api-v1"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "in": "path",
            "name": "process_id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Статус процесса",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Статус процесса",
                  "type": "object",
                  "properties": {
                    "process_id": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "PENDING",
                        "PARSING",
                        "READY",
                        "VERIFYING",
                        "COMPLETED",
                        "FINALIZED",
                        "ERROR",
                        "CANCELLED"
                      ]
                    },
                    "run_state": {
                      "type": "string"
                    },
                    "verification_status": {
                      "type": "string",
                      "enum": [
                        "PENDING",
                        "VERIFICATION_COMPLETED",
                        "PROTOCOL_FINALIZED"
                      ]
                    },
                    "stage": {
                      "type": "string"
                    },
                    "completed": {
                      "type": "integer"
                    },
                    "total": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "protocol_version": {
                      "type": "integer"
                    },
                    "sync_status": {
                      "type": "string",
                      "enum": [
                        "NOT_SENT",
                        "LOCAL_ONLY",
                        "SENT",
                        "PENDING_SYNC",
                        "SEND_FAILED",
                        "SEND_REFUSED"
                      ]
                    },
                    "scenario": {
                      "type": "string",
                      "enum": [
                        "FULL",
                        "PD_RD_ONLY",
                        "PD_ID_ONLY",
                        "RD_ID_ONLY",
                        "SINGLE_ONLY",
                        "PARTIALLY_LOADED",
                        "NO_DOCUMENTS"
                      ]
                    },
                    "upload_status": {
                      "type": "object"
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Причина отказа словами: что произошло и что сделать.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {}
                  },
                  "description": "Причина отказа словами: что произошло и что сделать."
                }
              }
            }
          },
          "403": {
            "description": "Причина отказа словами: что произошло и что сделать.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {}
                  },
                  "description": "Причина отказа словами: что произошло и что сделать."
                }
              }
            }
          },
          "422": {
            "description": "Причина отказа словами: что произошло и что сделать.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {}
                  },
                  "description": "Причина отказа словами: что произошло и что сделать."
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/processes/{process_id}": {
      "get": {
        "summary": "Протокол и результат процесса",
        "tags": [
          "api-v1"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "in": "path",
            "name": "process_id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/processes/{process_id}/cancel": {
      "post": {
        "summary": "Остановить проверку",
        "tags": [
          "api-v1"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "in": "path",
            "name": "process_id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/processes/{process_id}/decisions": {
      "post": {
        "summary": "Решение инспектора по кандидату",
        "tags": [
          "api-v1"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "finding_id",
                  "status",
                  "reason",
                  "expected_version"
                ],
                "properties": {
                  "finding_id": {
                    "type": "string",
                    "minLength": 1
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "CONFIRMED_VIOLATION",
                      "NEGATIVE_VERIFIED",
                      "CLARIFICATION_REQUIRED",
                      "CANDIDATE"
                    ]
                  },
                  "reason": {
                    "type": "string"
                  },
                  "reason_code": {
                    "type": "string"
                  },
                  "expected_version": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "author": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "in": "path",
            "name": "process_id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/processes/{process_id}/findings/{finding_id}/split": {
      "post": {
        "summary": "Разделить составной кандидат на атомарные findings",
        "tags": [
          "api-v1"
        ],
        "description": "ТЗ 9.3, п.2: каждая часть получает свои expected/actual и доказательства из карточки кандидата (индексы evidence) с координатами в двух стадиях; решение принимается по каждой части.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "reason",
                  "expected_version",
                  "parts"
                ],
                "properties": {
                  "reason": {
                    "type": "string"
                  },
                  "expected_version": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "parts": {
                    "type": "array",
                    "minItems": 2,
                    "items": {
                      "type": "object",
                      "required": [
                        "expected_value",
                        "actual_value",
                        "evidence_indexes"
                      ],
                      "properties": {
                        "expected_value": {
                          "type": "string"
                        },
                        "actual_value": {
                          "type": "string"
                        },
                        "evidence_indexes": {
                          "type": "array",
                          "minItems": 1,
                          "items": {
                            "type": "integer",
                            "minimum": 0
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "in": "path",
            "name": "process_id",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "finding_id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/processes/{process_id}/finalize": {
      "post": {
        "summary": "Финализировать протокол («Завершить»)",
        "tags": [
          "api-v1"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "author": {
                    "type": "string"
                  }
                },
                "nullable": true
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "in": "path",
            "name": "process_id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/processes/{process_id}/unfinalize": {
      "post": {
        "summary": "Отменить финализацию (администратор или супервизор)",
        "tags": [
          "api-v1"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "reason"
                ],
                "properties": {
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "in": "path",
            "name": "process_id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/processes/{process_id}/events": {
      "get": {
        "summary": "Журнал событий протокола",
        "tags": [
          "api-v1"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "in": "path",
            "name": "process_id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/processes/{process_id}/suspicions": {
      "get": {
        "summary": "Гипотезы свободного поиска (ТЗ 9.5)",
        "tags": [
          "api-v1"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "in": "path",
            "name": "process_id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/processes/{process_id}/suspicions/{suspicion_id}": {
      "post": {
        "summary": "Решение по гипотезе свободного поиска",
        "tags": [
          "api-v1"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "action"
                ],
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "promote",
                      "dismiss",
                      "confirm",
                      "reject"
                    ]
                  },
                  "comment": {
                    "type": "string"
                  },
                  "reason_code": {
                    "type": "string"
                  },
                  "evidence": {
                    "type": "array",
                    "items": {},
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "in": "path",
            "name": "process_id",
            "required": true
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "in": "path",
            "name": "suspicion_id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/processes/{process_id}/export": {
      "get": {
        "summary": "Протокол в JSON, XML, DOCX, PDF (или таблица проверок CSV)",
        "tags": [
          "api-v1"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "xml",
                "docx",
                "pdf",
                "csv"
              ],
              "default": "json"
            },
            "in": "query",
            "name": "format",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "in": "path",
            "name": "process_id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearer": []
    },
    {
      "cookie": []
    }
  ]
}